massive smog progress
