+++ Localization/de-DE/Mods.PathOfTerraria.Buffs.hjson
@@ -276,3 +276,8 @@ MosquitoTrappedDebuff: {
 	// DisplayName: Squeezed
 	// Description: Losing life
 }
+
+ToxicSmogDebuff: {
+	// DisplayName: Toxic Smog Debuff
+	// Description: Mods.PathOfTerraria.Buffs.ToxicSmogDebuff.Description
+}
