+++ Localization/ru-RU/Mods.PathOfTerraria.Buffs.hjson
@@ -280,3 +280,8 @@ MosquitoTrappedDebuff: {
 	// DisplayName: Squeezed
 	// Description: Losing life
 }
+
+ToxicSmogDebuff: {
+	// DisplayName: Toxic Smog Debuff
+	// Description: Mods.PathOfTerraria.Buffs.ToxicSmogDebuff.Description
+}
